Deprecating termText to conform with JLucene 2.3.2
git-svn-id: https://clucene.svn.sourceforge.net/svnroot/clucene/branches/lucene2_3_2@2809 20ef185c-fe11-0410-a618-ba9304b01011

+++ src/core/CLucene/document/DateTools.cpp
@@ -48,7 +48,7 @@ void DateTools::timeToString(const int64_t time, Resolution resolution, TCHAR* b
 
 	if (resolution == MILLISECOND_FORMAT) {
 		size_t len = strftime(abuf, DATETOOLS_BUFFER_SIZE, "%Y%m%d%H%M%S", ptm);
-		uint32_t ms = time % 1000;
+		uint32_t ms = static_cast<uint32_t>(time % 1000);
 		_snprintf(abuf + len, 4, "%03u", ms);
 	} else if (resolution == SECOND_FORMAT) {
 		strftime(abuf, DATETOOLS_BUFFER_SIZE, "%Y%m%d%H%M%S", ptm);
